Fix for race (#48)
* Fix for twin fetch race

* Properly fix multi queries to substrate

The idea here is that we use a worker pool to
do queries. With the pool we can control exactly
how much queries can be in flight. If too many are
requesting we limit the calls to 10 at a single moment
of time.

The worker will still try out the cache (again) before
it tries to query substrate. This way we make sure the
cache was not populated already by another worker in case
this worker was waiting for sometime to get a turn.

* remove commented out code